Refactor the PostgreSQL column comment support
Use the system catalogs directly instead of the information schema,
just as the other queries use.

Use Database#regclass_oid to correctly handle the PostgreSQL
search_path, and not assume the unqualified identifiers are in the
public schema.

Add the ability to support comments on other databases by defining
the appropriate function.